Use HornetQ transactions instead of custom guice scopes.
Original feature was for reliable event delivery, ensuring that we only send
events if the request successfully completes. To do this we added some
complicated Guice scopes so we could annotate objects that would exist only for
the request/job, and hold onto the queue of events to be sent at completion.

This showed up as a bit of a performance hit in load testing.

Instead we now leverage HornetQ transactions, events are queued but will not be
sent until commit is called on the session. We must make sure to call commit or
rollback on every request/job, but these hooks were already established for the
first cut of this. On rollback the events disappear and will not be sent out
even when the session is re-used the next time the thread is picked up.

Custom guice scopes are removed, and the event dispatcher is merged back
into the event sink class, as both objects are not really required anymore.
