+++ server/src/main/java/org/candlepin/audit/EventSink.java
@@ -22,6 +22,7 @@ import org.candlepin.model.Rules;
 import org.candlepin.model.activationkeys.ActivationKey;
 import org.candlepin.policy.js.compliance.ComplianceStatus;
 
+import java.util.List;
 import java.util.Set;
 
 /**
@@ -29,9 +30,13 @@ import java.util.Set;
  */
 public interface EventSink {
 
+    void initialize() throws Exception;
+
     void queueEvent(Event event);
 
     void sendEvents();
+
+    void rollback();
 
     void emitConsumerCreated(Consumer newConsumer);
 
@@ -52,4 +57,6 @@ public interface EventSink {
     void emitRulesDeleted(Rules rules);
 
     void emitCompliance(Consumer consumer, Set<Entitlement> entitlements, ComplianceStatus compliance);
+
+    List<QueueStatus> getQueueInfo();
 }
