--- server/src/main/java/org/candlepin/guice/PinsetterJobScoped.java
@@ -1,29 +1,0 @@
-/**
- * Copyright (c) 2009 - 2012 Red Hat, Inc.
- *
- * This software is licensed to you under the GNU General Public License,
- * version 2 (GPLv2). There is NO WARRANTY for this software, express or
- * implied, including the implied warranties of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. You should have received a copy of GPLv2
- * along with this software; if not, see
- * http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
- *
- * Red Hat trademarks are not licensed under GPLv2. No permission is
- * granted to use or replicate Red Hat trademarks that are incorporated
- * in this software or its documentation.
- */
-package org.candlepin.guice;
-
-import static java.lang.annotation.ElementType.*;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import com.google.inject.ScopeAnnotation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-/**
- * Items tagged with this annotation will be scoped to a PinsetterJob.
- */
-@Target({ TYPE, METHOD }) @Retention(RUNTIME) @ScopeAnnotation
-public @interface PinsetterJobScoped {}
