+++ server/src/main/java/org/candlepin/pinsetter/core/GuiceJobFactory.java
@@ -15,15 +15,12 @@
 package org.candlepin.pinsetter.core;
 
 import org.candlepin.guice.CandlepinSingletonScope;
-import org.candlepin.guice.SimpleScope;
 
 import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.persist.UnitOfWork;
 
 import org.quartz.Job;
-import org.quartz.JobExecutionContext;
-import org.quartz.JobExecutionException;
 import org.quartz.Scheduler;
 import org.quartz.SchedulerException;
 import org.quartz.spi.JobFactory;
@@ -31,13 +28,10 @@ import org.quartz.spi.TriggerFiredBundle;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.inject.Named;
-
 /**
  * GuiceJobFactory is a custom Quartz JobFactory implementation which
  * delegates job creation to the Guice injector.
- *
- * Courtesy https://github.com/brunojcm/guartz
+ * @version $Rev$
  */
 public class GuiceJobFactory implements JobFactory {
 
@@ -45,7 +39,6 @@ public class GuiceJobFactory implements JobFactory {
     private Injector injector;
     private CandlepinSingletonScope candlepinSingletonScope;
     private UnitOfWork unitOfWork;
-    @Inject @Named("PinsetterJobScope") private SimpleScope pinsetterJobScope;
 
     @Inject
     public GuiceJobFactory(Injector injector, CandlepinSingletonScope singletonScope,
@@ -55,71 +48,49 @@ public class GuiceJobFactory implements JobFactory {
         this.unitOfWork = unitOfWork;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public Job newJob(TriggerFiredBundle bundle, Scheduler scheduler)
         throws SchedulerException {
-        final Class<? extends Job> jobClass = bundle.getJobDetail().getJobClass();
-        Job decorator = new JobScopeDecorator(jobClass);
-        return decorator;
+        Class<Job> jobClass = (Class<Job>) bundle.getJobDetail().getJobClass();
+
+        boolean startedUow = startUnitOfWork();
+        candlepinSingletonScope.enter();
+        try {
+            return injector.getInstance(jobClass);
+        }
+        finally {
+            candlepinSingletonScope.exit();
+            if (startedUow) {
+                endUnitOfWork();
+            }
+        }
     }
 
-    private final class JobScopeDecorator implements Job {
-        private final Class<? extends Job> decoratedJobClass;
-        protected Job decorated;
-
-        private JobScopeDecorator(Class<? extends Job> jobClass) {
-            this.decoratedJobClass = jobClass;
-        }
-
-        @Override
-        public void execute(JobExecutionContext context) throws JobExecutionException {
-
-            // Enter custom scopes and start a unit of work just for the injection done
-            // when creating the job:
-            pinsetterJobScope.enter();
-            candlepinSingletonScope.enter();
-            boolean startedUow = startUnitOfWork();
+    protected boolean startUnitOfWork() {
+        if (unitOfWork != null) {
             try {
-                decorated = injector.getInstance(decoratedJobClass);
+                unitOfWork.begin();
+                return true;
             }
-            finally {
-                candlepinSingletonScope.exit();
-                if (startedUow) {
-                    endUnitOfWork();
-                }
+            catch (IllegalStateException e) {
+                log.debug("Already have an open unit of work");
+                return false;
             }
+        }
+        return false;
+    }
 
+    protected void endUnitOfWork() {
+        if (unitOfWork != null) {
             try {
-                decorated.execute(context);
+                unitOfWork.end();
             }
-            finally {
-                pinsetterJobScope.exit();
-            }
-        }
-
-        protected boolean startUnitOfWork() {
-            if (unitOfWork != null) {
-                try {
-                    unitOfWork.begin();
-                    return true;
-                }
-                catch (IllegalStateException e) {
-                    log.debug("Already have an open unit of work");
-                    return false;
-                }
-            }
-            return false;
-        }
-
-        protected void endUnitOfWork() {
-            if (unitOfWork != null) {
-                try {
-                    unitOfWork.end();
-                }
-                catch (IllegalStateException e) {
-                    log.debug("Unit of work is already closed, doing nothing");
-                    // If there is no active unit of work, there is no reason to close it
-                }
+            catch (IllegalStateException e) {
+                log.debug("Unit of work is already closed, doing nothing");
+                // If there is no active unit of work, there is no reason to close it
             }
         }
     }
